feat: load modes in own file
